Started a UTF-8 helper library.
We'll be using UTF-8 for all internal string storage. So might as well
have a library of useful functions for processing UTF-8 data. Right now
the only function I've added is one to find the index of the last UTF-8
boundary in a byte sequence. This function does not do much (if any)
UTF-8 validation, just gets the boundary index and returns it.