Implement RopeBuilder to build balanced Ropes from readers.
Since we can't get the size of data from a generic reader, it's hard to
build a balanced Rope. By clumping nodes together incrementally we'd get
a tree that was deeper on the left that on the right. Instead, the
RopeBuilder splits the data into a linear array of chunks, and at the
end clumps pairs of nodes together. There is the possibility that the
last node is never paired up, but that's okay, it doesn't introduce too
much imbalance.