Get an iterator over Rope chunks.
A chunk here is the data contained within a single Leaf node. The Chunks
iterator creates a stack of Inner nodes plus the leftmost Leaf node that
is within its range. On calling the next() method, the iterator gets the
current leftmost Leaf node if it exists and returns its data. If also
pops the next Inner node from the stack, explores down as far left as
possible, and places the leftmost Leaf node in the next_node field. This
goes on till the last Leaf node is reached, or we reach the end of the
range.