Create take_mut function to take and put back non-Copy data
In some cases, we need to take ownership of data from mutable references
and then put new data back, but Rust doesn't allow us to do that with
safe code. So I added a new function `take_mut` that takes a mutable
references to data, and a closure which takes owned data and returns
data of the same type, which will then be put back into the mutable
reference.