Add iterator over lines, and fix empty Rope cases.
I found a huge hole in my testing - I wasn't testing for the case of an
empty Rope!

Also added an iterator over lines in a Rope. Lines are delimited by the
newline character, '\n'. The iterator strips the terminating newline
from the Line before returning it. The iterator behaves line the Lines
iterator on Rust's internal String.