Remove take_mut to prevent un-debuggable crashes.
If the closure called by `take_mut` panics, that can potentially cause a
double free of the data that we took ownership of, since we don't really
move out of it. This causes the program to just "abort" without printing
a stack trace, and that's really hard to debug. So I figured that I
don't really need `take_mut` at the moment. Instead I'll do what I can
with a mutable reference.