New crate for buffer editing operations.
I'm splitting buffer editing logic into its separate crate so that I can
modularize it better, something I've struggled with in previous
iterations of the editor.

The idea is that all editing operations will go through a BufferView.
You can create a BufferView using the new_view() method on a buffer. A
buffer will stay alive as long as there is a BufferView for it.

Each BufferView contains a view_id for the particular buffer. This will
be used to get the per-view state, that is actually owned by the buffer.
This mechanism has been chosen to enable the buffer to modify each
view's state when another view modifies the buffer.